+++ CHANGELOG.md
@@ -27,6 +27,8 @@ If you love `detect-secrets`, please star our project on GitHub to show your sup
 
 <!--
 ### Unreleased
+#### :tada: New Features
+- Added Grafana tokens
 -->
 
 ### v1.4.0